Extract BsdPowerSource common base (NetBSD/OpenBSD dedup)
NetBsdPowerSource and OpenBsdPowerSource were identical except for
class names. Merge them into a single BsdPowerSource in the shared
unix hardware package, eliminating ~79 duplicated lines.
